requirements.txt
google-adk[database]==1.11.0
yfinance==0.2.56
psutil==5.9.5
litellm==1.66.3
google-generativeai==0.8.5
python-dotenv==1.1.0
PyMuPDF==1.24.9
python-docx==1.1.0
deprecated==1.2.14
# Google Calendar MCP dependencies
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
mcp==1.0.0
# Date/Time utilities
pytz==2023.3
